pyproject.toml
[project]
name = "second-brain"
version = "0"
description = "A second brain"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-frontmatter>=1.1.0",
    "ruff>=0.15.1",
]
